requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
